Changed dependencies to use root project libraries configuration. Renamed the gradle build file to conform to project standard for a module. Fixed findbugs reported issues. Removed unnecessary classes and instead use an anonymous class.
