--- hibernate-gradle-plugin/build.gradle
@@ -1,47 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * Copyright (c) 2013, Red Hat Inc. or third-party contributors as
- * indicated by the @author tags or express copyright attribution
- * statements applied by the authors.  All third-party contributions are
- * distributed under license by Red Hat Inc.
- *
- * This copyrighted material is made available to anyone wishing to use, modify,
- * copy, or redistribute it subject to the terms and conditions of the GNU
- * Lesser General Public License, as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
- * for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this distribution; if not, write to:
- * Free Software Foundation, Inc.
- * 51 Franklin Street, Fifth Floor
- * Boston, MA  02110-1301  USA
- */
-apply plugin: 'groovy'
-apply plugin: 'maven'
-apply plugin: 'java'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-   compile group: 'org.hibernate', name: 'hibernate-core', version: '4.2.2.Final'
-   compile group: 'org.hibernate.javax.persistence', name: 'hibernate-jpa-2.1-api', version: '1.0.0.Draft-16'
-   compile group: 'org.javassist', name: 'javassist', version: '3.15.0-GA'
-   compile gradleApi()
-   compile localGroovy()
-}
-
-/* Available for testing locally. */
-install {
-    repositories.mavenInstaller {
-        pom.groupId = 'org.hibernate'
-        pom.artifactId = 'hibernate-gradle-plugin'
-        pom.version = '0.1'
-    }
-}
