+++ hibernate-gradle-plugin/src/main/groovy/org/hibernate/bytecode/enhance/plugins/EnhancePlugin.java
@@ -31,34 +31,30 @@ import org.gradle.api.plugins.JavaBasePlugin;
 import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.plugins.JavaPlugin;
 import org.hibernate.bytecode.enhance.plugins.EnhanceTask;
-import org.hibernate.bytecode.enhance.plugins.EnhancePluginConvention;
 
 /**
  * This plugin will add Entity enhancement behaviour to the build lifecycle.
  * 
  * @author Jeremy Whiting
  */
-public class EnhancePlugin implements Plugin<Project> {
-
+@SuppressWarnings("serial")
+public class EnhancePlugin implements Plugin<Project>{
+	
 	public static final String ENHANCE_TASK_NAME = "enhance";
-	public static final String PLUGIN_CONVENTION_NAME = "enhance";
-	public static final String HAPPENS_BEFORE_ENHANCE_TASK_NAME = JavaPlugin.CLASSES_TASK_NAME;
 	public static final String HAPPENS_AFTER_ENHANCE_TASK_NAME = JavaPlugin.JAR_TASK_NAME;
 
 	public void apply(Project project) {
 		project.getLogger().debug( "Applying enhance plugin to project." );
-		project.getConvention().getPlugins().put( PLUGIN_CONVENTION_NAME, new EnhancePluginConvention() );
 		configureTask( project );
-		project.getLogger().debug( String.format( "DAG has been configured with enhance task dependent on [%s].", HAPPENS_BEFORE_ENHANCE_TASK_NAME ) );
+		project.getLogger().debug( String.format( "DAG has been configured with enhance task dependent on [%s].", JavaPlugin.CLASSES_TASK_NAME ) );
 	}
 
 	private void configureTask(Project project) {
 		EnhanceTask enhanceTask = project.getTasks().create( ENHANCE_TASK_NAME, EnhanceTask.class );
+		enhanceTask.setGroup(BasePlugin.BUILD_GROUP);
 		// connect up the task in the task dependency graph
-		final Configuration config = enhanceTask.getProject().getConfigurations().getByName( JavaPlugin.COMPILE_CONFIGURATION_NAME );
 		Task classesTask = project.getTasks().getByName( JavaPlugin.CLASSES_TASK_NAME );
 		enhanceTask.dependsOn( classesTask );
-		enhanceTask.mustRunAfter( HAPPENS_BEFORE_ENHANCE_TASK_NAME );
 
 		Task jarTask = project.getTasks().getByName( HAPPENS_AFTER_ENHANCE_TASK_NAME );
 		jarTask.dependsOn( enhanceTask );
