--- hibernate-gradle-plugin/src/main/groovy/org/hibernate/bytecode/enhance/plugins/EnhancePluginConvention.groovy
@@ -1,46 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * Copyright (c) 2013, Red Hat Inc. or third-party contributors as
- * indicated by the @author tags or express copyright attribution
- * statements applied by the authors.  All third-party contributions are
- * distributed under license by Red Hat Inc.
- *
- * This copyrighted material is made available to anyone wishing to use, modify,
- * copy, or redistribute it subject to the terms and conditions of the GNU
- * Lesser General Public License, as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
- * for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this distribution; if not, write to:
- * Free Software Foundation, Inc.
- * 51 Franklin Street, Fifth Floor
- * Boston, MA  02110-1301  USA
- */
-package org.hibernate.bytecode.enhance.plugins
-
-/**
-* Convention object for plugin.
-*/
-class EnhancePluginConvention {
-   
-    def contexts =  []
-   
-    public EnhancePluginConvention() {
-        contexts.add('org.hibernate.bytecode.enhance.contexts.AssociationReference')
-    }
-   
-    def add(final String fullyQualifiedClassName) {
-        contexts.add(fullyQualifiedClassName);
-    }
-    List getContexts() {
-        return contexts
-    }
-    def clear() {
-        contexts.clear();
-    }
-}
