+++ settings.gradle
@@ -10,6 +10,7 @@ include 'hibernate-proxool'
 
 include 'hibernate-ehcache'
 include 'hibernate-infinispan'
+include 'hibernate-gradle-plugin'
 include 'documentation'
 include 'release'
